python-examples/requirements.txt
grpcio==1.44.0
grpcio-tools==1.44.0
mypy-protobuf==3.2.0
